adding PageFactory & Suite
