+++ src/main/java/webdriver/factory/ChromeDriverCreator.java
@@ -1,0 +1,15 @@
+package webdriver.factory;
+
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+
+public class ChromeDriverCreator extends WebDriverCreator{
+    @Override
+    public WebDriver createWebDriver ()
+    {
+        System.setProperty("webdriver.chrome.driver", System.getProperty("user.dir") + "/drivers/chromedriver.exe");
+        return new ChromeDriver();
+    }
+}
