+++ src/main/java/webdriver/factory/WebDriverCreator.java
@@ -1,0 +1,9 @@
+package webdriver.factory;
+
+import org.openqa.selenium.WebDriver;
+
+public abstract class WebDriverCreator {
+
+    public abstract WebDriver createWebDriver();
+
+}
