+++ suites/suite-regression.xml
@@ -1,0 +1,31 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
+<suite name="suite-regression" verbose="1" thread-count="2" parallel="methods">
+    <parameter name="url" value="https://www.demoblaze.com/"/>
+    <test name="Executing tests in Firefox" >
+        <parameter name="browser" value="firefox"/>
+        <packages>
+            <package name="login" />
+            <package name="cart" />
+            <package name="product" />
+            <package name="purchase" />
+        </packages>
+    </test>
+    <test name="Executing tests in Chrome" >
+        <parameter name="browser" value="chrome"/>
+        <packages>
+            <package name="login" />
+            <package name="cart" />
+            <package name="product" />
+            <package name="purchase" />
+        </packages>
+    </test>
+    <test name="Executing tests in Edge" >
+        <parameter name="browser" value="edge"/>
+        <packages>
+            <package name="login" />
+            <package name="cart" />
+            <package name="product" />
+            <package name="purchase" />
+        </packages>
+    </test>
+</suite>
